Move lazy-load utils to its own lazyLoad.js file
